drop jersey helpers for annotation inheritance
This is for the case where a method overrides a method of a superclass / interface
and we want to merge the annotations of the overriden method with the annotations
of the current method (on return-value, parameters).

Also,
- fix a bug in checking bean field modifiers
- for bean properties merge annotations on field and getter only

Drop setter annotations because it is ambiguous which annotations to use. The
right thing seems to be to use the parameter annotations, but users will expect
to use the method (i.e. the void return value) annotations.